# openapi.yaml
openapi: 3.0.0
paths:
...
  /residents/{residentUid}/elimination-records:
    get:
      operationId: getEliminationRecords
      parameters:
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema:
            type: string
      responses:
        "200":
          description: 排泄記録一覧の取得に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EliminationRecordListResponseDto"
      security:
        - JWT-auth: []
      summary: 利用者の排泄記録一覧を取得
      tags: &a7
        - elimination-records
    post:
      operationId: createEliminationRecord
      parameters:
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EliminationRecordCreateInputDto"
      responses:
        "201":
          description: 排泄記録の作成に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EliminationRecordDto"
      security:
        - JWT-auth: []
      summary: 排泄記録を作成
      tags: *a7
  /residents/{residentUid}/elimination-records/{uid}:
    patch:
      operationId: updateEliminationRecord
      parameters:
        - name: uid
          required: true
          in: path
          description: 排泄記録UID
          schema:
            type: string
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema: {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EliminationRecordUpdateInputDto"
      responses:
        "200":
          description: 排泄記録の更新に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EliminationRecordDto"
      security:
        - JWT-auth: []
      summary: 排泄記録を更新
      tags: *a7
    delete:
      operationId: deleteEliminationRecord
      parameters:
        - name: uid
          required: true
          in: path
          description: 排泄記録UID
          schema:
            type: string
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema: {}
      responses:
        "200":
          description: 排泄記録の削除に成功
      security:
        - JWT-auth: []
      summary: 排泄記録を削除
      tags: *a7
  /residents/{residentUid}/elimination-records/{uid}/transcription:
    get:
      operationId: getEliminationRecordTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: 排泄記録UID
          schema:
            type: string
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema: {}
      responses:
        "200":
          description: 排泄記録の文字起こし取得に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionDto"
      security:
        - JWT-auth: []
      summary: 排泄記録の文字起こしを取得
      tags: *a7
    patch:
      operationId: appendEliminationRecordTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: 排泄記録UID
          schema:
            type: string
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema: {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranscriptionInputDto"
      responses:
        "200":
          description: 排泄記録の文字起こし追記に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionDto"
      security:
        - JWT-auth: []
      summary: 排泄記録の文字起こしを追記
      tags: *a7
    put:
      operationId: updateEliminationRecordTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: 排泄記録UID
          schema:
            type: string
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema: {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranscriptionInputDto"
      responses:
        "200":
          description: 排泄記録の文字起こし置換に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionDto"
      security:
        - JWT-auth: []
      summary: 排泄記録の文字起こしを置換
      tags: *a7
    delete:
      operationId: deleteEliminationRecordTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: 排泄記録UID
          schema:
            type: string
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema: {}
      responses:
        "200":
          description: 排泄記録の文字起こし削除に成功
      security:
        - JWT-auth: []
      summary: 排泄記録の文字起こしを削除
      tags: *a7
  /residents/{residentUid}/elimination-records/{uid}/extract:
    post:
      operationId: extractEliminationRecord
      parameters:
        - name: uid
          required: true
          in: path
          description: 排泄記録UID
          schema:
            type: string
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema: {}
      responses:
        "200":
          description: 排泄記録からの情報抽出に成功
          content:
            application/json:
              schema:
                type: string
      security:
        - JWT-auth: []
      summary: 排泄記録から情報を抽出
      tags: *a7
...
  /assessments/{uid}/transcription:
    get:
      operationId: getTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: アセスメントUID
          schema:
            type: string
      responses:
        "200":
          description: アセスメントの文字起こし取得に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionDto"
      security:
        - JWT-auth: []
      summary: アセスメントの文字起こしを取得
      tags: *a10
    patch:
      operationId: appendTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: アセスメントUID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranscriptionInputDto"
      responses:
        "200":
          description: アセスメントの文字起こし追記に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionDto"
      security:
        - JWT-auth: []
      summary: アセスメントの文字起こしを追記
      tags: *a10
    put:
      operationId: updateTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: アセスメントUID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranscriptionInputDto"
      responses:
        "200":
          description: アセスメントの文字起こし置換に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionDto"
      security:
        - JWT-auth: []
      summary: アセスメントの文字起こしを置換
      tags: *a10
    delete:
      operationId: deleteTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: アセスメントUID
          schema:
            type: string
      responses:
        "200":
          description: アセスメントの文字起こし削除に成功
      security:
        - JWT-auth: []
      summary: アセスメントの文字起こしを削除
      tags: *a10
...
info:
  title: Nagara Care API
  description: ""
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    JWT-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Enter JWT token
      in: header
  schemas:
...
    TranscriptionDto:
      type: object
      properties:
        transcription:
          type: string
          description: 文字起こしテキスト
      required:
        - transcription
    TranscriptionInputDto:
      type: object
      properties:
        transcription:
          type: string
          description: 文字起こしテキスト
      required:
        - transcription
...
    EliminationRecordDto:
      type: object
      properties:
        uid:
          type: string
          description: UID
        tenantUid:
          type: string
          description: テナントUID
        caregiverUid:
          type: string
          description: 介護者UID
        residentUid:
          type: string
          description: 利用者UID
        recordedAt:
          format: date-time
          type: string
          description: 記録時刻
        notes:
          type: string
          description: メモ
        eliminationMethod:
          type: string
          description: 排泄方法
        hasFeces:
          type: boolean
          description: 便の有無
        fecalIncontinence:
          type: boolean
          description: 便失禁の有無
        fecesAppearance:
          type: string
          description: 便の性状
        fecesVolume:
          type: number
          description: 便の量（g）
        hasUrine:
          type: boolean
          description: 尿の有無
        urinaryIncontinence:
          type: boolean
          description: 尿失禁の有無
        urineAppearance:
          type: string
          description: 尿の性状
        urineVolume:
          type: number
          description: 尿量（ml）
      required:
        - uid
        - tenantUid
        - caregiverUid
        - residentUid
        - recordedAt
        - notes
        - eliminationMethod
        - hasFeces
        - fecalIncontinence
        - fecesAppearance
        - fecesVolume
        - hasUrine
        - urinaryIncontinence
        - urineAppearance
        - urineVolume
    EliminationRecordListResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/EliminationRecordDto"
        total:
          type: number
          description: 総件数
      required:
        - items
        - total
    EliminationRecordCreateInputDto:
      type: object
      properties:
        recordedAt:
          format: date-time
          type: string
          description: 記録時刻
          example: 2024-03-20T12:00:00Z
        notes:
          type: string
          description: メモ
        eliminationMethod:
          type: string
          description: 排泄方法
          example: トイレ
        hasFeces:
          type: boolean
          description: 便の有無
          example: true
        fecalIncontinence:
          type: boolean
          description: 便失禁の有無
          example: false
        fecesAppearance:
          type: string
          description: 便の性状
          example: 普通便
        fecesVolume:
          type: number
          description: 便の量（g）
          minimum: 0
          example: 100
        hasUrine:
          type: boolean
          description: 尿の有無
          example: true
        urinaryIncontinence:
          type: boolean
          description: 尿失禁の有無
          example: false
        urineAppearance:
          type: string
          description: 尿の性状
          example: 普通
        urineVolume:
          type: number
          description: 尿量（ml）
          minimum: 0
          example: 200
      required:
        - recordedAt
        - eliminationMethod
        - hasFeces
        - hasUrine
    EliminationRecordUpdateInputDto:
      type: object
      properties:
        recordedAt:
          format: date-time
          type: string
          description: 記録時刻
          example: 2024-03-20T12:00:00Z
        notes:
          type: string
          description: メモ
        eliminationMethod:
          type: string
          description: 排泄方法
          example: トイレ
        hasFeces:
          type: boolean
          description: 便の有無
          example: true
        fecalIncontinence:
          type: boolean
          description: 便失禁の有無
          example: false
        fecesAppearance:
          type: string
          description: 便の性状
          example: 普通便
        fecesVolume:
          type: number
          description: 便の量（g）
          minimum: 0
          example: 100
        hasUrine:
          type: boolean
          description: 尿の有無
          example: true
        urinaryIncontinence:
          type: boolean
          description: 尿失禁の有無
          example: false
        urineAppearance:
          type: string
          description: 尿の性状
          example: 普通
        urineVolume:
          type: number
          description: 尿量（ml）
          minimum: 0
          example: 200
      required:
        - recordedAt
        - eliminationMethod
        - hasFeces
        - hasUrine
...